Change to 'watchr' module.
This also updates primeBuffet() and rebuild() to
accept an optional rootPath argument. If a path
is passed then the cache will only be primed/rebuilt
from the point in the directory structure down.

Uses the new path-specific rebuild to respond to
new directory events.

All tests pass in OSX now.  Needs Linux testing.